D3/P1: Optimize memory to O(1) when counting bits
Refactor most common bit per column resolution to use a generator
instead of an array which optimizes theorical memory consumption
for that part of the algo to O(1).

To be honest that's kinda pointless given the number of bits was static,
but it's just for fun. Memory consumption could be greatly optimized if
a similar strategy was used to read the input file though, perhaps later.